update package name update RELEASE.md
