+++ RELEASE.md
@@ -1,6 +1,13 @@
-# Release 1.0.4
+# Release 1.0.6
+
+- Update app package name
+
+# Release 1.0.5
 
 - Update Android minimum API Level required for the application to run to 21
+
+# Release 1.0.4
+
 - Update app launcher icons to include adaptive and legacy launcher icons
 
 # Release 1.0.3
