+++ maps-app/build.gradle
@@ -5,7 +5,7 @@ android {
 
 
     defaultConfig {
-        applicationId "com.esri.android.mapsapp"
+        applicationId "com.esri.arcgisruntime.opensourceapps.mapsapp"
         minSdkVersion rootProject.minSdkVersion
         targetSdkVersion rootProject.targetSdkVersion
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
