+++ maps-app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.esri.android.mapsapp">
+          package="com.esri.arcgisruntime.opensourceapps.mapsapp">
 
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION"/>
@@ -16,7 +16,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
         <activity
-            android:name="com.esri.android.mapsapp.MapsAppActivity"
+            android:name="com.esri.arcgisruntime.opensourceapps.mapsapp.MapsAppActivity"
             android:configChanges="orientation|screenSize"
             android:windowSoftInputMode="stateAlwaysHidden"
             android:label="@string/app_name">
@@ -27,7 +27,7 @@
         </activity>
 
         <activity
-            android:name="com.esri.android.mapsapp.account.SignInActivity"
+            android:name="com.esri.arcgisruntime.opensourceapps.mapsapp.account.SignInActivity"
             android:label="@string/title_sign_in_activity"
             android:launchMode="standard"
             android:uiOptions="splitActionBarWhenNarrow"
